Rename User on Employee, now Employee is an odinary class not an abstract class
