+++ src/main/java/com/codecool/Dao/LoginDao.java
@@ -1,4 +1,10 @@
 package com.codecool.Dao;
 
-public class LoginDao {
+import com.codecool.Model.Employee;
+
+import java.util.List;
+
+public interface LoginDao {
+    List<Employee> getUsers();
+    Employee getUserByLogin(String login);
 }
