+++ src/main/java/com/codecool/Dao/UserDao.java
@@ -1,11 +1,11 @@
 package com.codecool.Dao;
 
-import com.codecool.Model.User;
+import com.codecool.Model.Employee;
 
 import java.util.List;
 
 public interface UserDao {
-    List<User> getAll();
+    List<Employee> getAll();
     void delete(int ids);
     void add(String name, String surname, String mail, String loginID);
     void edit(int id);
