--- src/main/java/com/codecool/Model/Admin.java
@@ -1,4 +1,0 @@
-package com.codecool.Model;
-
-public class Admin {
-}
