+++ src/main/java/com/codecool/Model/Employee.java
@@ -1,4 +1,31 @@
 package com.codecool.Model;
 
 public class Employee {
+    int id;
+    String name;
+    String surname;
+    String email;
+
+    public Employee(int id, String name, String surname, String email) {
+        this.id = id;
+        this.name = name;
+        this.surname = surname;
+        this.email = email;
+    }
+
+    public int getID() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getSurname() {
+        return surname;
+    }
+
+    public String getEmail() {
+        return email;
+    }
 }
