--- src/main/java/com/codecool/Model/Mentor.java
@@ -1,4 +1,0 @@
-package com.codecool.Model;
-
-public class Mentor {
-}
