--- src/main/java/com/codecool/Model/User.java
@@ -1,31 +1,0 @@
-package com.codecool.Model;
-
-public abstract class User {
-    int id;
-    String name;
-    String surname;
-    String email;
-
-    public User(int id, String name, String surname, String email) {
-        this.id = id;
-        this.name = name;
-        this.surname = surname;
-        this.email = email;
-    }
-
-    public int getID() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-}
